+++ front/components/assistant_builder/DataSourceSelectionSection.tsx
@@ -7,9 +7,9 @@ import {
   TrashIcon,
 } from "@dust-tt/sparkle";
 
-import type { AssistantBuilderDataSourceConfiguration } from "@app/components/assistant_builder/AssistantBuilder";
 import { CONNECTOR_PROVIDER_TO_RESOURCE_NAME } from "@app/components/assistant_builder/shared";
 import { EmptyCallToAction } from "@app/components/EmptyCallToAction";
+import type { AssistantBuilderDataSourceConfiguration } from "@app/lib/api/assistant_builder/types";
 import { CONNECTOR_CONFIGURATIONS } from "@app/lib/connector_providers";
 import { getDisplayNameForDataSource } from "@app/lib/data_sources";
