+++ front/components/assistant_builder/DustAppSelectionSection.tsx
@@ -6,8 +6,8 @@ import {
 } from "@dust-tt/sparkle";
 import { Transition } from "@headlessui/react";
 
-import type { AssistantBuilderDustAppConfiguration } from "@app/components/assistant_builder/AssistantBuilder";
 import { EmptyCallToAction } from "@app/components/EmptyCallToAction";
+import type { AssistantBuilderDustAppConfiguration } from "@app/lib/api/assistant_builder/types";
 
 export default function DustAppSelectionSection({
   show,
