+++ front/lib/client/tables_query.ts
@@ -1,0 +1,7 @@
+import type { TablesQueryConfigurationType } from "@dust-tt/types";
+
+export function tableKey(
+  table: TablesQueryConfigurationType["tables"][number]
+) {
+  return `${table.workspaceId}/${table.dataSourceId}/${table.tableId}`;
+}
